Move ShaderStages to pipeline::shader
