+++ vulkano/src/command_buffer/synced/commands.rs
@@ -26,7 +26,6 @@ use crate::command_buffer::ImageUninitializedSafe;
 use crate::command_buffer::SecondaryCommandBuffer;
 use crate::command_buffer::SubpassContents;
 use crate::descriptor_set::layout::DescriptorDescTy;
-use crate::descriptor_set::layout::ShaderStages;
 use crate::descriptor_set::DescriptorSet;
 use crate::format::ClearValue;
 use crate::image::ImageAccess;
@@ -35,6 +34,7 @@ use crate::pipeline::depth_stencil::DynamicStencilValue;
 use crate::pipeline::depth_stencil::StencilFaceFlags;
 use crate::pipeline::input_assembly::IndexType;
 use crate::pipeline::layout::PipelineLayout;
+use crate::pipeline::shader::ShaderStages;
 use crate::pipeline::viewport::Scissor;
 use crate::pipeline::viewport::Viewport;
 use crate::pipeline::ComputePipelineAbstract;
