+++ vulkano/src/descriptor_set/layout/sys.rs
@@ -163,8 +163,8 @@ mod tests {
     use crate::descriptor_set::layout::DescriptorDesc;
     use crate::descriptor_set::layout::DescriptorDescTy;
     use crate::descriptor_set::layout::DescriptorSetLayout;
-    use crate::descriptor_set::layout::ShaderStages;
     use crate::descriptor_set::pool::DescriptorsCount;
+    use crate::pipeline::shader::ShaderStages;
     use std::iter;
 
     #[test]
